+++ src/test/java/seedu/rainyDay/command/AddCommandTest.java
@@ -1,0 +1,59 @@
+package seedu.rainyDay.command;
+
+import org.junit.After;
+import org.junit.Before;
+import org.junit.jupiter.api.Test;
+import seedu.rainyDay.data.FinancialReport;
+import seedu.rainyDay.data.FinancialStatement;
+
+import java.io.ByteArrayOutputStream;
+import java.io.PrintStream;
+import java.util.ArrayList;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
+public class AddCommandTest {
+    ArrayList<FinancialStatement> statements = new ArrayList<>();
+    FinancialReport financialReport = new FinancialReport(statements);
+
+    private final ByteArrayOutputStream outContent = new ByteArrayOutputStream();
+    private final ByteArrayOutputStream errContent = new ByteArrayOutputStream();
+    private final PrintStream originalOut = System.out;
+    private final PrintStream originalErr = System.err;
+
+    @Before
+    public void setUpStreams() {
+        System.setOut(new PrintStream(outContent));
+        System.setErr(new PrintStream(errContent));
+    }
+
+    @After
+    public void restoreStreams() {
+        System.setOut(originalOut);
+        System.setErr(originalErr);
+    }
+
+    @Test
+    public void execute_singleStatement_statementInformation() {
+        setUpStreams();
+        AddCommand addCommand = new AddCommand("Ipad", "out", 120);
+        addCommand.setData(financialReport);
+        addCommand.execute();
+        String expectedAddStatement = "Done! Added: out for Ipad, -$120" + System.lineSeparator();
+        assertEquals(expectedAddStatement, outContent.toString());
+        restoreStreams();
+    }
+
+    @Test
+    public void execute_multipleStatements_statementsInformation() {
+        setUpStreams();
+        AddCommand addCommand = new AddCommand("angpao", "in", 3000);
+        addCommand.setData(financialReport);
+        addCommand.execute();
+        String expectedAddStatement = "Done! Added: in for angpao, +$3000" + System.lineSeparator();
+        assertEquals(expectedAddStatement, outContent.toString());
+        restoreStreams();
+        //Command.addFinancialStatement("Ipad", "out", 120);
+        //TODO: assertequal() to check if size is 2, and if the 2 tasks are correctly stored
+    }
+}
